+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # WIP
 
 - Stable sorting of sequences of various types #470
+- Correctly restore `:max-tx` from storage
 
 # 1.6.5 - May 3, 2024